6 follow-up improvements: editor selectability, gestures, sub-pixel precision
#1 Editor selectability:
- Joystick is no longer attached to the edit-touch listener (its
  activation zone is large and dragging it doesn't help layout).
- row_top_buttons LinearLayout dropped — Esc / E / F now individual
  editable widgets at TopEnd anchors with marginEnd 16/72/128.
- Sneak / Sprint already individuals from the previous fan-layout
  commit; selection visual still uses View.foreground yellow ring.

#2 UI mode editor:
- column_ui_buttons LinearLayout dropped — five UI-mode buttons
  (LMB / RMB / Q / Shift / Esc) become individuals at BottomStart
  anchors stacked with vertical margins 20 / 104 / 188 / 256 / 324.

#3 Sprint trigger symmetric:
- ENGAGE and DISENGAGE now both 1.2x rim radius (no hysteresis).

#4 Sub-pixel look-delta precision (10x):
- LookPadView multiplies finger deltas by SUBPIXEL_SCALE (10) before
  the residual+truncate step; wire i16 now carries tenths-of-pixel.
- PC HandleLookDelta divides by 10 before applying CameraCurve.
- CameraCurve.Apply takes float now (existing tests still pass —
  int args implicit-convert; 40/40 tests pass).
- Effect: a 0.3px finger move emits delta 3 (= 0.3 px), preserving
  fine micro-aim that previously truncated to zero.

#5 In-game lookpad gestures:
- onSingleTapConfirmed → MOUSE_LEFT click (down + 50ms + up).
- Double-tap-and-hold → MOUSE_LEFT down on second tap's DOWN, up on
  second tap's UP/CANCEL. Camera delta tracking continues during the
  hold so the user can drag while LMB is held.

#6 UI-mode lookpad gestures:
- Single tap = MOUSE_LEFT click (same as in-game).
- Quick double tap = MOUSE_RIGHT click.

Wire (LookPadView) routes to gestures based on the mode StateFlow
mirrored from server, so a STATE_CHANGE on the wire also retargets
the lookpad's behavior.

+++ pc/McController.Server/Input/CameraCurve.cs
@@ -27,10 +27,10 @@ public sealed class CameraCurve
         _config = config;
     }
 
-    public (int sdx, int sdy) Apply(int rawDx, int rawDy)
+    public (int sdx, int sdy) Apply(float rawDx, float rawDy)
     {
         var cam = _config.Camera;
-        var speed = MathF.Sqrt((float)(rawDx * rawDx + rawDy * rawDy));
+        var speed = MathF.Sqrt(rawDx * rawDx + rawDy * rawDy);
 
         float accelMul = cam.CurveType switch
         {
